+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Pin alphatwirl explicitly to 0.25.5, PR #130
+
 ## [0.18.2] - 2020-07-01
 ### Added
 - Add option to BinnedDataframe to apply weights to data like MC, PR #127 [@BenKrikler](httsp://github.com/benkrikler)