+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+
+## [0.19.0] - 2020-08-19
 ### Added
 - Put a tarball with book-keeping data in output directory, PR #131 [@BenKrikler](httsp://github.com/benkrikler)
 - Add Dask and Parsl backends from Coffea and support execution configuration via YAML, PR #129 [@BenKrikler](https://github.com/benkrikler) and [@asnaylor](https://github.com/asnaylor)
 
 ### Fixed
 - Pin alphatwirl explicitly to 0.25.5, PR #130
+
+### Removed
+- Removed help module from carpenter and moved to fast-flow, PR #131 [@BenKrikler](https://github.com/benkrikler)
 
 ## [0.18.2] - 2020-07-01
 ### Added